requirements.txt
blynklib==0.2.6
pyserial==3.5
RPi.GPIO==0.7.0
